sign up/in customer v1.0

+++ src/main/java/com/javalearn/camerastore/controller/admin/CustomerController.java
@@ -30,7 +30,7 @@ public class CustomerController {
     }
 
     @DeleteMapping("/customer")
-    public void deleteCustomer(@RequestParam int id)
+    public void deleteCustomer(@RequestParam Long id)
     {
         customerService.deleteCustomer(id);
     }
